Extract IntensityMatcher from block worker
